Exclude 'staging' env from bootstrapping keyring
The 'staging' development environment via its make target [1] is already
equipped with setting up the appropriate 'staging' keyring. In turn this
staging keyring gives access to the staging version of keyring. with the
inclusion of the keyring installation in 'make install-rpm' [2], this
step in OpenQA's side is no longer necessary.

[1]: https://github.com/freedomofpress/securedrop-workstation/pull/1500/files#diff-76ed074a9305c04054cdebb9e9aad2d818052b07091de1f20cad0bbac34ffb52R32-R33
[2]: https://github.com/freedomofpress/securedrop-workstation/pull/1500/files#diff-76ed074a9305c04054cdebb9e9aad2d818052b07091de1f20cad0bbac34ffb52R59-R61