Treat 'prod-qa' env as 'prod'
The workstation installation was failing because 'prod-qa' was being set
in the workstation's config.json file. But there prod-qa doesn't
actually carry any meaning. This commit sets the env as 'prod' in the
case where it's "prod-qa", but then instructs OpenQA to replace the
prod repo with a prod-qa one.